Thread ids are longs, not ints.

+++ dyninstAPI/h/BPatch.h
@@ -221,7 +221,7 @@ public:
     void registerNormalExit(process *proc, int exitcode);
     void registerSignalExit(process *proc, int signalnum);
 
-    void registerThreadExit(process *proc, int tid);
+    void registerThreadExit(process *proc, long tid);
 
     void registerProcess(BPatch_process *process, int pid=0);
     void unRegisterProcess(int pid);
